backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Image handling
Pillow==10.1.0
cloudinary==1.36.0

# HTTP client
httpx==0.25.2

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Optional: for better performance
redis==5.0.1

# Geospatial utilities
shapely==2.0.2
pyproj==3.6.1

# Date/time utilities
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.2.0

# CORS
fastapi-cors==0.0.6